fix: handle user decline on transaction signing, fix install from source, and use ruff (#91)
* fix: handle user decline

* docs: show how to install from source

* fix: safe fixes

* refactor(Account): remove `SignatureError`; streamline logic

---------

+++ scripts/build.py
@@ -8,10 +8,19 @@ PLUGIN_MANIFEST_FOLDER = project.path / "ape_safe" / "manifests"
 
 
 @click.command()
-def cli():
+@click.option("--override", is_flag=True)
+def cli(override):
     for version in project.dependencies["safe-contracts"]:
         dependency = project.dependencies["safe-contracts"][version]
         plugin_manifest_path = PLUGIN_MANIFEST_FOLDER / f"safe-{version}.json"
+
+        if plugin_manifest_path.is_file() and not override:
+            click.echo(f"Version '{version}' already exists, skipping.'")
+            continue
+
+        # OK to delete - either doesn't exist or override set.
+        plugin_manifest_path.unlink(missing_ok=True)
+
         click.echo(
             f"cp $HOME/{dependency.manifest_path.relative_to(Path.home())}"
             f" ./{plugin_manifest_path.relative_to(Path.cwd())}"
